requirements.txt
flask~=3.0.2
SQLAlchemy~=2.0.27
pandas~=2.2.1
pytz~=2024.1
pillow~=10.2.0
numpy~=1.26.4
flask_sqlalchemy
fpdf
num2words
